Stringify nulled $json variable
For no apparent reason the json_decode() in load() started to throw an error recently. 
It's true, as I set the values for the first time for a new user the $json returns null. But this should not throw an error (which it did: "json_decode() expects parameter 1 to be string, null given"). This fixes this.